backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-socketio==5.10.0
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
python-dotenv==1.0.0
requests==2.31.0
pillow==10.1.0
numpy==1.26.2
giphy_client==1.0.0
python-multipart==0.0.6 